Allow using 4MB mmcblk1p13 bpsw partition for droid4-kexecboot
We now have buildroot based image using kexecboot GUI at:

https://github.com/tmlind/buildroot

With flashable images at:

https://github.com/tmlind/droid4-kexecboot

Let's add support for booting to the buildroot rootfs on
mmcblk1p13 bpsw partition. This partition does not seem to be
used by the stock firmware based on reflashing the modem
firmware.

Note that to get busybox init started on buildroot rootfs, we
need to start with init='/bin/busybox init' probably because of
selinux on the stock firmware.

Let's also stop trying to install anything by default as we
probably want to move to using droid4-kexecboot instead of
kexec scripts on the distro. The stock v3.0.8 kernel is quite
useless so best to isolate all the hacks to the 4MB partition.