Add proper argument handling and use it
The product name may not always be p2a_maserati depending on the
device. It can be read from the device with:

$ fastboot getvar product

Let's add proper argument handling and use it to prepare for adding
other support for other products like mz619. Additionally mz619 uses
serialno tag as the first tag that we need to add support for.

This causes no intended changes to the output file.

Signed-off-by: Tony Lindgren <tony@atomide.com>